requirements.txt
langchain-google-genai==1.0.10
langchain-community==0.2.16
langchain==0.2.16
google==3.0.0
python-dotenv==1.0.1
